gateway/requirements.txt
awsiotsdk==1.21.0
